Ensure canvas keeps keyboard focus
Canvas is focusable. Before, it would lose focus if you click away from it, but regain focus if you click back on it.

Lately (at least in Chrome), the canvas been losing focus permanently when you click away from it. That is, clicking back on the canvas does not cause it to resume receiving keyboard events.

Therefore, do this automatically and programmatically using canvas.onblur, which seems to fix the issue.