Add experimental version
The experimental version is an evolution of the design and
implementation. It is very much in flux.

It has some notable changes and additions, but is still missing some of
the basics (such as most of the editors, APIs, interpreter, undo/redo,
etc.) that are present in the stable version. Much of the internals
have been rewritten.

Changes:
- resolution: 192x128 --> 240x160
- colors: 16 --> 64 (6bpp)
- fps: 30 --> 60
- memory map
- font

Additions:
- terminal colors (fg/bg)
- syntax highlighting via simple regular expressions